+++ go/kinds/composable/histogram/panel/x/histogram_types_gen.go
@@ -159,8 +159,8 @@ type ScaleDistribution string
 
 // TODO docs
 type ScaleDistributionConfig struct {
-	LinearThreshold *int `json:"linearThreshold,omitempty"`
-	Log             *int `json:"log,omitempty"`
+	LinearThreshold *float32 `json:"linearThreshold,omitempty"`
+	Log             *float32 `json:"log,omitempty"`
 
 	// TODO docs
 	Type ScaleDistribution `json:"type"`
@@ -187,7 +187,7 @@ type VizLegendOptions struct {
 	ShowLegend bool            `json:"showLegend"`
 	SortBy     *string         `json:"sortBy,omitempty"`
 	SortDesc   *bool           `json:"sortDesc,omitempty"`
-	Width      *int32          `json:"width,omitempty"`
+	Width      *float32        `json:"width,omitempty"`
 }
 
 // TODO docs
